fix(db): declare rule_id bigint in the hypertable swap, as 001_init.sql does
migrateToTimescaleDB builds logs_hypertable from a hardcoded column list and
renames it over logs_partitioned. That runs on fresh installs, and its
rule_id was integer while 001_init.sql declares bigint — so every new
install ended up with the narrow column the schema says it should not have,
permanently: ADD COLUMN IF NOT EXISTS never repairs a type. ModSecurity rule
ids above 2147483647 then fail to insert with "integer out of range".

Verified against a throwaway fresh database: after the swap the column is
bigint and stores 4294967295; the same insert on a pre-fix fresh install
still errors. The two column lists are otherwise identical — same 36
columns, same names, same order, and the only other differences are
`NOT NULL DEFAULT x` vs `DEFAULT x NOT NULL` clause order on id, timestamp
and created_at, which PostgreSQL treats identically.

Retro-widening installs that already ran the swap is deliberately not here:
it is a full table rewrite that has to decompress every chunk first.

Refs #297

+++ api/internal/database/migration.go
@@ -1776,7 +1776,11 @@ func (db *DB) migrateToTimescaleDB() {
 			http_x_forwarded_for text,
 			severity log_severity,
 			error_message text,
-			rule_id integer,
+			-- bigint, not integer: ModSecurity rule ids exceed int4 (2147483647),
+			-- 001_init.sql declares logs_partitioned.rule_id bigint, and this table
+			-- is renamed over it on fresh installs — a narrower type here would
+			-- stick forever, since ADD COLUMN IF NOT EXISTS never repairs a type (#297)
+			rule_id bigint,
 			rule_message text,
 			rule_severity text,
 			rule_data text,
